console-start function added in new class startApp
